Update for multiple app('captcha')->display() on the same page (#57)
* Update for multiple app('captcha')->display() on the same page

registerin the service provider with singleton instead of bind will return always the same object per page load. with this methos is possible to insert a control variable in the class and use it to 

* Upda for multiple noCaptcha on the same page

ading this var control will permit to request the api only once. this will work with the singleton register on te serviceprovider file.

* Update NoCaptcha.php

* Update NoCaptcha.php